ref: 사진 삭제 API 통합 (단건, 다중)
