+++ src/main/kotlin/com/yapp2app/photo/application/command/PhotoImageCommand.kt
@@ -27,8 +27,6 @@ data class GetFavoritePhotosCommand(
     val sortOrder: SortOrder = SortOrder.DESC,
 )
 
-data class DeletePhotoCommand(val userId: Long, val photoId: Long)
-
 data class DeletePhotosCommand(val userId: Long, val photoIds: List<Long>)
 
 data class UpdatePhotoCommand(val userId: Long, val photoId: Long, val memo: String?)
