+++ src/main/kotlin/com/yapp2app/photo/application/port/PhotoImageRepositoryPort.kt
@@ -17,7 +17,6 @@ interface PhotoImageRepositoryPort {
 
     fun listOwnedFavoritePhotos(userId: Long, offset: Int, limit: Int, sortOrder: SortOrder): List<PhotoImage>
 
-    fun deleteOwnedPhoto(userId: Long, photoId: Long): PhotoImage?
     fun deleteOwnedPhotos(userId: Long, photoIds: List<Long>): List<PhotoImage>
 
     fun getOwnedPhoto(userId: Long, photoId: Long): PhotoImage?
