+++ src/main/kotlin/com/yapp2app/photo/infra/client/PhotoMediaClient.kt
@@ -1,7 +1,6 @@
 package com.yapp2app.photo.infra.client
 
 import com.yapp2app.media.application.command.ConfirmMediaUploadedCommand
-import com.yapp2app.media.application.command.DeleteMediaCommand
 import com.yapp2app.media.application.command.DeleteMediasCommand
 import com.yapp2app.media.application.command.GetMediaStorageInfosCommand
 import com.yapp2app.media.application.command.GetMediasCommand
@@ -54,10 +53,6 @@ class PhotoMediaClient(
                 contentType = it.contentType,
             )
         }
-    }
-
-    override fun deleteMedia(ownerId: Long, mediaId: Long) {
-        deleteMediaUseCase.execute(DeleteMediaCommand(ownerId, mediaId))
     }
 
     override fun deleteMedias(ownerId: Long, mediaIds: List<Long>) {
