+++ src/main/kotlin/com/yapp2app/photo/infra/persist/PhotoImageRepositoryAdapter.kt
@@ -36,15 +36,6 @@ class PhotoImageRepositoryAdapter(
         sortOrder: SortOrder,
     ): List<PhotoImage> = queryRepository.findOwnedFavoritePhotos(userId, offset, limit, sortOrder)
 
-    override fun deleteOwnedPhoto(userId: Long, photoId: Long): PhotoImage? {
-        val photo = jpaRepository.findByUserIdAndId(userId, photoId)
-            ?: return null
-
-        jpaRepository.delete(photo)
-
-        return photo
-    }
-
     override fun deleteOwnedPhotos(userId: Long, photoIds: List<Long>): List<PhotoImage> {
         val photos = jpaRepository.findAllByUserIdAndIdIn(userId, photoIds)
