+++ src/test/kotlin/com/yapp2app/e2e/photo/image/DeletePhotosE2ETest.kt
@@ -19,7 +19,7 @@ import org.springframework.test.context.ActiveProfiles
  * fileName       : DeletePhotosE2ETest
  * author         : koo
  * date           : 2026. 1. 8.
- * description    : DELETE /api/photos (bulk delete) E2E 테스트
+ * description    : DELETE /api/photos E2E 테스트 (단건 삭제 및 다건 삭제)
  */
 @ActiveProfiles("test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -39,6 +39,25 @@ class DeletePhotosE2ETest : PhotoImageE2ETestBase() {
         val (user, token) = createTestUserAndToken()
         testUser = user
         accessToken = token
+    }
+
+    @Test
+    @DisplayName("단건 삭제 성공")
+    fun givenSinglePhotoId_whenDeletePhotos_thenReturnsSuccess() {
+        // given
+        val media = createMedia(ownerId = testUser.id!!, status = MediaStatus.UPLOADED)
+        val photo = createPhotoImage(userId = testUser.id!!, mediaId = media.id!!)
+
+        // when & then
+        RestAssured.given()
+            .contentType(ContentType.JSON)
+            .header("Authorization", "Bearer $accessToken")
+            .body(DeletePhotosRequest(photoIds = listOf(photo.id!!)))
+            .`when`()
+            .delete("/api/photos")
+            .then()
+            .statusCode(HttpStatus.OK.value())
+            .body("resultCode", equalTo(ResultCode.SUCCESS.code))
     }
 
     @Test
@@ -167,5 +186,90 @@ class DeletePhotosE2ETest : PhotoImageE2ETestBase() {
             .then()
             .statusCode(HttpStatus.OK.value())
             .body("resultCode", equalTo(ResultCode.SUCCESS.code))
+    }
+
+    @Test
+    @DisplayName("존재하지 않는 사진 ID만 있을 때 성공을 반환한다")
+    fun givenNonExistentPhotoIdOnly_whenDeletePhotos_thenReturnsSuccess() {
+        // given
+        val nonExistentPhotoId = 999999L
+
+        // when & then
+        RestAssured.given()
+            .contentType(ContentType.JSON)
+            .header("Authorization", "Bearer $accessToken")
+            .body(DeletePhotosRequest(photoIds = listOf(nonExistentPhotoId)))
+            .`when`()
+            .delete("/api/photos")
+            .then()
+            .statusCode(HttpStatus.OK.value())
+            .body("resultCode", equalTo(ResultCode.SUCCESS.code))
+    }
+
+    @Test
+    @DisplayName("다른 사용자의 사진 ID만 있을 때 성공을 반환한다")
+    fun givenOtherUserPhotoIdOnly_whenDeletePhotos_thenReturnsSuccess() {
+        // given
+        val (otherUser, _) = createTestUserAndToken(email = "other2@example.com")
+        val media = createMedia(ownerId = otherUser.id!!, status = MediaStatus.UPLOADED)
+        val otherUserPhoto = createPhotoImage(userId = otherUser.id, mediaId = media.id!!)
+
+        // when & then
+        RestAssured.given()
+            .contentType(ContentType.JSON)
+            .header("Authorization", "Bearer $accessToken")
+            .body(DeletePhotosRequest(photoIds = listOf(otherUserPhoto.id!!)))
+            .`when`()
+            .delete("/api/photos")
+            .then()
+            .statusCode(HttpStatus.OK.value())
+            .body("resultCode", equalTo(ResultCode.SUCCESS.code))
+    }
+
+    @Test
+    @DisplayName("폴더에 포함된 사진 단건 삭제 성공")
+    fun givenPhotoInFolder_whenDeletePhotos_thenReturnsSuccess() {
+        // given
+        val folder = createFolder(userId = testUser.id!!)
+        val media = createMedia(ownerId = testUser.id!!, status = MediaStatus.UPLOADED)
+        val photo = createPhotoImage(userId = testUser.id!!, mediaId = media.id!!, folderId = folder.id)
+
+        // when & then
+        RestAssured.given()
+            .contentType(ContentType.JSON)
+            .header("Authorization", "Bearer $accessToken")
+            .body(DeletePhotosRequest(photoIds = listOf(photo.id!!)))
+            .`when`()
+            .delete("/api/photos")
+            .then()
+            .statusCode(HttpStatus.OK.value())
+            .body("resultCode", equalTo(ResultCode.SUCCESS.code))
+    }
+
+    @Test
+    @DisplayName("즐겨찾기된 사진 삭제 시 즐겨찾기도 함께 삭제된다")
+    fun givenFavoritePhoto_whenDeletePhotos_thenFavoriteAlsoDeleted() {
+        // given
+        val media = createMedia(ownerId = testUser.id!!, status = MediaStatus.UPLOADED)
+        val photo = createFavoritePhotoImage(userId = testUser.id!!, mediaId = media.id!!)
+
+        // when
+        RestAssured.given()
+            .contentType(ContentType.JSON)
+            .header("Authorization", "Bearer $accessToken")
+            .body(DeletePhotosRequest(photoIds = listOf(photo.id!!)))
+            .`when`()
+            .delete("/api/photos")
+            .then()
+            .statusCode(HttpStatus.OK.value())
+
+        // then
+        RestAssured.given()
+            .header("Authorization", "Bearer $accessToken")
+            .`when`()
+            .get("/api/photos/favorite/summary")
+            .then()
+            .statusCode(HttpStatus.OK.value())
+            .body("data.totalCount", equalTo(0))
     }
 }
